feat: add targetname base class to ms fgd
There are many MS specific entities which are trying to implement the Targetname base class in the MS .fgd, but there is no such base class by that name; it exists in halflife.fgd, and is likely the result of folks copy pasting over when creating new entities like I did. This commit adds a Targetname base class to the .fgd, which should seamlessly populate the entities referencing it with the appropriate key/value pairs.

This prevents the "Warning: undefined base class 'Targetname'" error as seen in J.A.C.K. on loading the ms1.4.fgd

Note that since this change affects a large number of entities at once, it should be extensively tested prior to merging.