feat: add trigger base class to ms fgd
This commit adds a Trigger base class to the .fgd, which should seamlessly populate the entities referencing it with the appropriate key/value pairs.

This prevents the "Warning: undefined base class 'Trigger'" error as seen in J.A.C.K. on loading the ms1.4.fgd

Note that since this change affects more than one entity, it should be extensively tested prior to merging.