Fgd fixes (#14)
Added numerous base classes which other classes in the FGD expected to
be there but weren't, resulting in errors as seen in J.A.C.K. when
loading the FGD.

Also removed a duplicate giantboar def which was causing an error.

Switched from `base(...Renderless)` to `base(...MSRenderless)` for
`ms_func_playerclip` to match other MS specific entities.